AMBARI-24822. Convert singletons to an enum
