+++ ambari-server/src/main/java/org/apache/ambari/server/controller/internal/TrimmingStrategy.java
@@ -18,7 +18,44 @@
 
 package org.apache.ambari.server.controller.internal;
 
-public interface TrimmingStrategy {
-  String trim(String stringToTrim);
+/**
+ * Various strategies for trimming properties during blueprint deployment.
+ */
+public enum TrimmingStrategy {
+
+  DEFAULT {
+    @Override
+    public String trim(String stringToTrim) {
+      return stringToTrim.trim();
+    }
+  },
+
+  DIRECTORIES {
+    @Override
+    public String trim(String stringToTrim) {
+      return stringToTrim.replaceAll("\\s*,+\\s*", ",").trim();
+    }
+  },
+
+  PASSWORD {
+    @Override
+    public String trim(String stringToTrim) {
+      return stringToTrim;
+    }
+  },
+
+  DELETE_SPACES_AT_END {
+    @Override
+    public String trim(String stringToTrim) {
+      if (" ".equals(stringToTrim)) {
+        return stringToTrim;
+      }
+      return stringToTrim.replaceAll("\\s+$", "");
+    }
+  }
+  ;
+
+  public abstract String trim(String stringToTrim);
+
 }
